Mocked services
Update documentation, rename test services to mocked services, remove
any built files
