+++ README.md
@@ -1,73 +1,58 @@
-# eLife Dashboard
+# [eLife Dashboard](https://github.com/elifesciences/elife-dashboard/)
 
-This is the eLife Dashboard application and API. 
+This is the eLife Dashboard application and API.
 
-[Click here for frontend documentation](README-Frontend.md).
+This document covers the setup of the dashboard application and API. [Click here for frontend documentation](README-Frontend.md) which includes instructions for a setup with no external dependencies.
 
----
+## Table of Contents
 
-# Requirements:
+
+* [Requirements](#requirements)
+* [Quick start](#quick-start)
+* [Tests](#tests)
+* [Documentation](#documentation)
+* [License](#license)
+
+
+
+## Requirements:
+
 
 * [Scheduler](https://github.com/elifesciences/elife-article-scheduler)  
-* Working python development environment
-* Postgres should be installed (this is not required for the test services)
+* [Python](https://www.python.org/) 
+* [Postgres](https://www.postgresql.org/)
 
----
 
+## Quick Start
 
 The instructions below are for Mac OS and may need to be adapted for other platforms
 
-## Create and use a virtual environment
+* Create virtual environment, you will only need to do this once. In the project folder run:
 
-create the environment (once):
-
-(in project folder)
 ```bash
 virtualenv venv
 ```
-
-to activate the virtual environment at any point:
+* Activate the virtual environment
 
 ```bash
 source venv/bin/activate
 ```
-
-to deactivate:
-
-```bash
-deactivate
-```
-
-### Install Python pre-requisites
-
-(in project folder)
-(with active virtual environment as above)
-
-if you only need to run the test services:
-```bash
-pip install -r service-test-requirements.txt
-
-```
-
-else
+* Install dependencies (This will only succeed if postgres binaries are installed)
 
 ```bash
 pip install -r requirements.txt
 ```
-This will only succeed if postgres binaries are installed
-
-
-### Running test services
-
-After setup above and with virtual environment active:
+* run application
 
 ```bash
-python test_services.py
+python runserver.py
 ```
 
-The server will launch and a console message will provide the base URL, typically http://localhost:8008
 
-### Running Python Unit tests
+
+
+## Tests
+
 
 The dashboard unit tests depend on database access. The settings used in the tests will be in `settings_test.py`.
 
@@ -82,24 +67,47 @@ cd /srv/elife-dashboard
 
 #### Pycharm:
 
-Create a new Run/Debug Configuration:
-Add a new Python test > Unittests configuration
-In the section
-Unittests:
-    Test: Script
-    Script: the path select the path .../tests/test_file_you want_to_test.py
-Environment:
-    Project: elife-dashboard
-    Environment variables: CONFIG_FILE=test
-    Python interpreter: (the virtualenv at your dashboard project)
-    Working directory: [your path to]/elife-dashboard
+* Create a new Run/Debug Configuration
+* Add a new Python test > Unittests configuration
+* In the section
+ * Unittests:
+     * Test: Script
+    	* Script: the path select the path .../tests/test_file_you want_to_test.py
+ * Environment:
+     * Project: elife-dashboard
+     * Environment variables: CONFIG_FILE=test
+     * Python interpreter: (the virtualenv at your dashboard project)
+     * Working directory: [your path to]/elife-dashboard
+ * Add content roots to PYTHONPATH (checked)
+ * Add source roots to PYTHONPATH (checked)
 
-Add content roots to PYTHONPATH (checked)
-Add source roots to PYTHONPATH (checked)
+
+
+
+## Documentation
+
+### Creating, Activating and Deactivating virtual environment
+
+**Create environment**
+
+```bash
+virtualenv venv
+```
+
+**Activate environment**
+
+```bash
+source venv/bin/activate
+```
+
+**Deactivate environment**
+
+```bash
+deactivate
+```
+
 
 
 ## License
 
 MIT Licensed. See [LICENSE](LICENSE)
-
-npm install
