+++ assets/js/main.js
@@ -1,13 +1,9 @@
 (function () {
-    var options = {
-        debug: true
-    };
     var config = require('config')();
     var schedule = require('schedule')(config);
     var current = require('current')(config);
     var detail = require('detail')(config);
     var scheduled = require('scheduled')(config);
-
     schedule.init();
     current.init();
     detail.init();
